% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcm-package.R
\docType{package}
\name{lcm-package}
\alias{lcm}
\alias{lcm-package}
\title{lcm: Nonparametric Longitudinal Causal Mediation}
\description{
Nonparametric longitudinal causal mediation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nt-williams/lcm}
  \item Report bugs at \url{https://github.com/nt-williams/lcm/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Iván Díaz \email{ild2005@med.cornell.edu} (\href{https://orcid.org/0000-0001-9056-2047}{ORCID}) [copyright holder]
}

}
\keyword{internal}
