% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcm.R
\name{lcm}
\alias{lcm}
\title{Causal Mediation with Longitudinal Mediators, Treatments, and Confounders}
\usage{
lcm(data, a_prime, a_star, Npsem, lrnrs, V)
}
\arguments{
\item{data}{A data frame}

\item{a_prime}{Value of a', 0 or 1.}

\item{a_star}{Value of a*, 0 or 1.}

\item{Npsem}{An `lcm_Npsem` object mapping observed variables to the assumed NPSEM}

\item{lrnrs}{An object inhering from `Lrnr_base` from the sl3 package}

\item{V}{The number of folds for cross-fitting}
}
\value{
An object of class `lcm`
}
\description{
Causal Mediation with Longitudinal Mediators, Treatments, and Confounders
}
\examples{
Np <- lcm_Npsem$new(
    L = list(c("L_1"), c("L_2")),
    A = c("A_1", "A_2"),
    Z = list(c("Z_1"), c("Z_2")),
    M = c("M_1", "M_2"),
    Y = "Y"
)

lcm(sim, 0, 0, Np, sl3::Lrnr_glm_fast$new(), 5)
}
